requirements.txt
# Packages required to run files in this repository and utils which I like.
Pygments
PySocks
Sphinx
Unidecode
distribute
ipython
nltk
networkx
numpydoc
srtmerge
sympy
# This seems really promising for parsing.
#pyparsing
termcolor
virtualenv
yolk
